pyproject.toml
[project]
name = "genshin-quest-to-md"
version = "0.1.0"
description = "把原神任务剧情 JSON 转为 Markdown 的小脚本。"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
json2md = "json_to_md.cli:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
